Develop branch

+++ src/main/java/com/example/pokemons/repositories/TypeRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface TypeRepository extends JpaRepository<Type, Integer> {
+public interface TypeRepository extends JpaRepository<Type, Long> {
 
     Optional<Type> findByName(String name);
 }
